+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Video playback using the Dynamic Adaptive Streaming over HTTP protocol.
+
 ## [0.19.0-RC5] 2017-12-01
 
 ### Fixed